向下取整

+++ isnet_pro/Inference2.py
@@ -488,7 +488,7 @@ def IS_inference_mask(img_mode,dataset_path,output_dir,ui_set_aim_bacground_rgb,
                             img_bacground = io.imread(bc_path)
                     elif(i%len(bc_list)==0):
                         if(i>=curbg):
-                            curbg=curbg+len(im_list)/len(bc_list)
+                            curbg=curbg+int(len(im_list)/len(bc_list))
                             bc_path = bc_list[bgindex]
                             img_bacground = io.imread(bc_path)
                             bgindex=bgindex+1
